requirements.txt
psycopg2-binary
pyodbc
pymysql
